+++ packages/yarnpkg-libui/sources/misc/renderForm.tsx
@@ -9,14 +9,13 @@ type InferProps<T extends React.ComponentType> = T extends React.ComponentType<i
 
 export type SubmitInjectedComponent<T, C extends React.ComponentType = React.ComponentType> = React.ComponentType<InferProps<C> & { useSubmit: (value: T) => void }>;
 
-// TODO: make the streams required in the next major so that people don't forget to pass them
 export type RenderFormOptions = {
-  stdin?: Readable;
-  stdout?: Writable;
-  stderr?: Writable;
+  stdin: Readable;
+  stdout: Writable;
+  stderr: Writable;
 };
 
-export async function renderForm<T, C extends React.ComponentType = React.ComponentType>(UserComponent: SubmitInjectedComponent<T, C>, props: InferProps<C>, {stdin, stdout, stderr}: RenderFormOptions = {}) {
+export async function renderForm<T, C extends React.ComponentType = React.ComponentType>(UserComponent: SubmitInjectedComponent<T, C>, props: InferProps<C>, {stdin, stdout, stderr}: RenderFormOptions) {
   let returnedValue: T | undefined;
 
   const useSubmit = (value: T) => {